% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{clean_df}
\alias{clean_df}
\title{Clean a given expression data frame, i.e. remove all genes with too many missing values and all genes with 0 variance}
\usage{
clean_df(my_df)
}
\arguments{
\item{my_df}{data frame holding expression data with column=sample, row=gene}
}
\value{
cleaned data frame
}
\description{
Clean a given expression data frame, i.e. remove all genes with too many missing values and all genes with 0 variance
}
