% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{enough_power}
\alias{enough_power}
\title{Tests if a gene set is large enough to provide for enough distinguishable gene-combinations in a given number of (boostrap runs)}
\usage{
enough_power(samplesize, my_set_size, nsims)
}
\arguments{
\item{samplesize}{The number of genes in one sample to be drawn.}

\item{my_set_size}{The size of the pool to draw from (size of this MSigDB gene set)}

\item{nsims}{Number of bootstrap runs}
}
\description{
Tests if a gene set is large enough to provide for enough distinguishable gene-combinations in a given number of (boostrap runs)
}
\keyword{Internal}
