% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{estimate_laplace_parameters}
\alias{estimate_laplace_parameters}
\title{Fits a Laplace distribution to the given expression data of treatment and control.}
\usage{
estimate_laplace_parameters(
  treatment_df = data.frame(),
  control_df = data.frame(),
  GOI,
  geneset_collection,
  output_dir = getwd(),
  plot_ECDF = TRUE,
  plot_histo = TRUE,
  histo_breaks = 100
)
}
\arguments{
\item{treatment_df}{Expression matrix of treatment condition (gene sets associated more strongly with GOI in this condition will appear on the right-hand side of the plot).}

\item{control_df}{Expression matrix of treatment condition (gene sets associated more strongly with GOI in this condition will appear on the right-hand side of the plot).}

\item{GOI}{A string specifying the GOI gene symbol.}

\item{geneset_collection}{A list of lists. Each list holds gene symbols of a specific gene set. Can be generated with \code{\link{get_msigdb_genesets}}}

\item{output_dir}{If the Laplace distribution fitting results should not only be returned as stats but also plotted, provide the output path for the plots. Default: working directory.}

\item{plot_ECDF}{Should the ECDF of the Laplace fitting be plotted? Default: true.}

\item{plot_histo}{Should the histogram of the Laplace fitting be plotted? Default: true.}

\item{histo_breaks}{How many breaks should the histogram have? Default: 100.}
}
\value{
Returns a named list with "location" (the location estimate / mean of the fitted Laplace distribution), "scale" (the scale / standard deviation of the fitted Laplace distribution) and "ks_test" (the result of the Kolmogorov Smirnov test).
}
\description{
Fits a Laplace distribution to the given expression data of treatment and control.
}
