% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_differential_results}
\alias{get_differential_results}
\title{Differential analysis of two given GeneCOCOA results for conditions treatment and control. Returns a dataframe with "neglog10_p_ratio" indicating direction & strength of change in association between a gene set and the GOI (>0: more strongly associated in treatment, <0: more strongly associated in control), and "differential_p" the corresponding significance of the change in association.}
\usage{
get_differential_results(
  treatment_res,
  control_res,
  GOI,
  geneset_collection,
  include_low_power = TRUE,
  laplace_parameters = "Default",
  treatment_df = data.frame(),
  control_df = data.frame()
)
}
\arguments{
\item{treatment_res}{Output of \code{\link{get_stats}} ("treatment" condition, right side of the plot).}

\item{control_res}{Output of \code{\link{get_stats}} ("control" condition, left side of the plot).}

\item{GOI}{A string specifying the GOI gene symbol.}

\item{geneset_collection}{A list of lists. Each list holds gene symbols of a specific gene set. Can be generated with \code{\link{get_msigdb_genesets}}}

\item{include_low_power}{Should gene sets which (for combinatorial reasons) could only be bootrstrapped <1000 times be included?}

\item{laplace_parameters}{Which Laplace distribution should be fitted to the data in order to derive p-values for the differential association values? Options: "Default": use precomputed Laplace-values (location=0.0076, scale=1.6614), "Laplace": location=0, scale=1, "Estimate": estimate parameters from given empirical data - cave: this will take longer than the other parameters!}

\item{treatment_df}{Only needed for laplace_parameters="Estimate". Expression matrix of treatment condition (gene sets associated more strongly with GOI in this condition will appear on the right-hand side of the plot).}

\item{control_df}{Only needed for laplace_parameters="Estimate". Expression matrix of treatment condition (gene sets associated more strongly with GOI in this condition will appear on the right-hand side of the plot).}
}
\value{
Returns differential GeneCOCOA results: a data frame with with columns "geneset", "p_control", "p_adj_control", "p_disease". "p_adj_disease", "p_ratio", "log10_p_ratio", "neglog10_p_ratio", "differential_p"
}
\description{
Differential analysis of two given GeneCOCOA results for conditions treatment and control. Returns a dataframe with "neglog10_p_ratio" indicating direction & strength of change in association between a gene set and the GOI (>0: more strongly associated in treatment, <0: more strongly associated in control), and "differential_p" the corresponding significance of the change in association.
}
