% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_expr_info.df}
\alias{get_expr_info.df}
\title{Takes a data frame holding expression data with rows = genes, columns = samples and transforms it to Gene-COCOA input format.}
\usage{
get_expr_info.df(expr_df, GOI, clean = TRUE)
}
\arguments{
\item{clean}{}
}
\description{
Takes a data frame holding expression data with rows = genes, columns = samples and transforms it to Gene-COCOA input format.
}
\keyword{Internal}
