% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_expr_info.file}
\alias{get_expr_info.file}
\title{Takes a path to a csv / tsv of expression values (rows=genes, columns=samples) and transforms it to Gene-COCOA input format.}
\usage{
get_expr_info.file(expr_file, GOI, clean = TRUE)
}
\arguments{
\item{clean}{}
}
\description{
Takes a path to a csv / tsv of expression values (rows=genes, columns=samples) and transforms it to Gene-COCOA input format.
}
\keyword{Internal}
