% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_geom_mean_cor}
\alias{get_geom_mean_cor}
\title{Computes the geometric mean coexpression between a given gene set and the GOI.}
\usage{
get_geom_mean_cor(GOI, GOI_expr, my_t.tpms)
}
\arguments{
\item{my_t.tpms}{}
}
\description{
Computes the geometric mean coexpression between a given gene set and the GOI.
}
