% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_geom_mean_expr}
\alias{get_geom_mean_expr}
\title{Computes the geometric mean of the expression of a given gene set.}
\usage{
get_geom_mean_expr(t.tpm.genes_in_set)
}
\arguments{
\item{t.tpm.genes_in_set}{}
}
\description{
Computes the geometric mean of the expression of a given gene set.
}
\keyword{Internal}
