% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_intercept}
\alias{get_intercept}
\title{Computes the intercept of each regression model}
\usage{
get_intercept(GOI, GOI_expr, my_t.tpms)
}
\arguments{
\item{my_t.tpms}{}
}
\description{
Computes the intercept of each regression model
}
\keyword{Internal}
