% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_mean_cor}
\alias{get_mean_cor}
\title{Computes the arithmetic mean of the coexpression between each gene in a set and the GOI.}
\usage{
get_mean_cor(GOI, GOI_expr, my_t.tpms)
}
\arguments{
\item{my_t.tpms}{}
}
\description{
Computes the arithmetic mean of the coexpression between each gene in a set and the GOI.
}
\keyword{Internal}
