% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_msigdb_genesets}
\alias{get_msigdb_genesets}
\title{Return a collection of curated gene sets from the MSigDB.}
\usage{
get_msigdb_genesets(genesets = "", remove_prefix = TRUE)
}
\arguments{
\item{remove_prefix}{Bool. Should the collection prefix (e.g. "HALLMARK_") be removed from the gene set names. (FALSE might result in cluttered plots.)}

\item{geneset}{Name of gene set. Can be "HALLMARK", "POSITIONAL", "KEGG", "PID", "REACTOME", "WP" (for WikiPathways), "GOBP" (for GO: Biological Process), "GOCC" (for GO: Cellular Component), "GOMF" (for GO: Molecular Function), "ONCO" (for oncogenic signature). If you are unsure which collection of gene sets to choose for comparison, "HALLMARK" is recommended.}
}
\value{
A list of named lists. Each sublist is named by a gene set, and its contents are the symbols of all genes in this set as strings.
}
\description{
Return a collection of curated gene sets from the MSigDB.
}
\examples{
\dontrun{
# assign gene set collection to variable and pass to get_stats
all_hallmark_gs <- get_msigdb_genesets("HALLMARK")
res <- get_stats(geneset_collection=all_hallmark_gs, GOI="PDGFD",  ...)

# call directly in get_stats()
res <- get_stats(geneset_collection=get_msigdb_genesets("HALLMARK"), ...)
}
}
