% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_neg_log10}
\alias{get_neg_log10}
\title{Returns -log10 of given p-value list}
\usage{
get_neg_log10(p_value_list)
}
\arguments{
\item{p_value_list}{}
}
\description{
Returns -log10 of given p-value list
}
\keyword{Internal}
