% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{get_stats}
\alias{get_stats}
\title{Main function of Gene-COCOA: performs statistical analysis.}
\usage{
get_stats(
  geneset_list,
  GOI,
  GOI_expr,
  expr_df,
  samplesize = ncol((expr_df)/10),
  nsims = 1000
)
}
\arguments{
\item{geneset_list}{A list of lists. Each list holds gene symbols of a specific gene set. Can be generated with \code{\link{get_msigdb_genesets}}}

\item{GOI}{A string specifying the GOI gene symbol.}

\item{GOI_expr}{The GOI expression per sample. Should be generated with get_expr_info(expr, GOI)$GOI_expr.}

\item{expr_df}{Complete expression data. Should be generated with get_expr_info(expr, GOI)$expr_df.}

\item{samplesize}{How many predictors should be in each regression model? Should approximate (number of samples)/ 10.}

\item{nsims}{Bootstrapping: how many simulations should be run? At least 1k recommended.}
}
\value{
A list of data frames holding the results of the Gene-COCOA analysis.
}
\description{
Main function of Gene-COCOA: performs statistical analysis.
}
\examples{
\dontrun{
# let CAD_disease: data frame holding expression data in coronary artery disease with rows=genes, columns=samples

expr_info <- get_expr_info(expr=CAD_disease, GOI="PDGFD")
res <- get_stats(geneset_list=get_msigigdb_genesets("HALLMARK"), GOI="PDGFD", GOI_expr=expr_info$GOI_expr, expr_df=expr_info$expr_df)
}
}
