% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot.basic_p_ranking}
\alias{plot.basic_p_ranking}
\title{Plots p-value ranking (x-axis=adjusted -log10(p), y= gene set names, ranked).
This function works but \code{\link{plot.volcano}} is more informative.}
\usage{
\method{plot}{basic_p_ranking}(my_stats, filepath)
}
\arguments{
\item{my_stats}{Output of \code{\link{get_stats}}}

\item{filepath}{Path + name of png to save If not provided, only the ggplot will be returned by this function.}
}
\value{
Returns a ggplot which can be stored in a variable and modified. If an absolute file name (path + file name) is provided, the ggplot with additionally be saved to a png.
}
\description{
Plots p-value ranking (x-axis=adjusted -log10(p), y= gene set names, ranked).
This function works but \code{\link{plot.volcano}} is more informative.
}
