% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot.multiple_p_plots}
\alias{plot.multiple_p_plots}
\title{Function to plot multiple (long) p-value rankings. Obsolete.}
\usage{
\method{plot}{multiple_p_plots}(subdf)
}
\description{
Function to plot multiple (long) p-value rankings. Obsolete.
}
\keyword{Internal}
