% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot.p_vs_geom_mean_cor}
\alias{plot.p_vs_geom_mean_cor}
\title{Returns a ggplot with x=-log10(p.adj), y=mean coexpression of the gene set with the GOI.
This function works but \code{\link{plot.volcano}} is more informative.}
\usage{
\method{plot}{p_vs_geom_mean_cor}(
  mystats,
  sig_label_cutoff = 0.05,
  sig_colour = "dodgerblue4",
  remove_outliers = FALSE,
  filepath
)
}
\arguments{
\item{mystats}{Output of \code{\link{get_stats}}.}

\item{filepath}{Path + name of png to save If not provided, only the ggplot will be returned by this function.}
}
\value{
Returns a ggplot which can be stored in a variable and modified. If an absolute file name (path + file name) is provided, the ggplot with additionally be saved to a png.
}
\description{
Returns a ggplot with x=-log10(p.adj), y=mean coexpression of the gene set with the GOI.
This function works but \code{\link{plot.volcano}} is more informative.
}
