% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot.volcano}
\alias{plot.volcano}
\title{Recommended visualisation of Gene-COCOA results.
Produces a ggplot with x=(GOI coexpression with gene set)/(GOI coexpression with random set), y=-log10(p.adj).
Dot size and alpha reflect the (geometric) mean expression of all genes in a respective gene set.}
\usage{
\method{plot}{volcano}(
  mystats,
  sig_label_cutoff = 0.01,
  sig_colour = "dodgerblue4",
  remove_outliers = FALSE,
  filepath = ""
)
}
\arguments{
\item{mystats}{Output of \code{\link{get_stats}}.}

\item{sig_label_cutoff}{At which p.adj-level should points be labelled? Default=0.01.}

\item{sig_colour}{Which colour should significant dots have?}

\item{remove_outliers}{Should outliers (in terms of x-axis and y-axis) be removed. Default is FALSE; you should stick with this unless you get very cluttered plots.}

\item{filepath}{Path + name of png to save If not provided, only the ggplot will be returned by this function.}
}
\value{
Returns a ggplot which can be stored in a variable and modified. If an absolute file name (path + file name) is provided, the ggplot with additionally be saved to a png.
}
\description{
Recommended visualisation of Gene-COCOA results.
Produces a ggplot with x=(GOI coexpression with gene set)/(GOI coexpression with random set), y=-log10(p.adj).
Dot size and alpha reflect the (geometric) mean expression of all genes in a respective gene set.
}
\examples{
\dontrun{
expr_info <- get_expr_info(expr=CAD_disease, GOI="PDGFD")
res <- get_stats(geneset_list=get_msigigdb_genesets("HALLMARK"), GOI="PDGFD", GOI_expr=expr_info$GOI_expr, expr_df=expr_info$expr_df)
# show GeneCOCOA ggplot in "Plots" window
plot_volcano(res)

# assign GeneCOCOA ggplot to variable
p <- plot_volcano(res)

# assign GeneCOCOA ggplot to variable AND save GeneCOCOA ggplot to file
p <- plot_volcano(res, filepath="/path/to/my/GeneCOCOA.volcano.png")
}
}
