% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot_control_vs_treatment}
\alias{plot_control_vs_treatment}
\title{Comparative visualisation of the GeneCOCOA results for a GOI in two different conditions.
Produces a ggplot divering bar plot with x=-log10(p.adj) and y= top significant terms in chosen condition}
\usage{
plot_control_vs_treatment(
  control_res,
  treatment_res,
  include_low_power = TRUE,
  control_label = "control",
  treatment_label = "treatment",
  control_col = "#00000000",
  treatment_col = "brown4",
  sort_by = "treatment",
  topN = 10,
  filepath = ""
)
}
\arguments{
\item{control_res}{Output of \code{\link{get_stats}} ("control" condition, left side of the plot).}

\item{treatment_res}{Output of \code{\link{get_stats}} ("treatment" condition, right side of the plot).}

\item{include_low_power}{Should gene sets which (for combinatorial reasons) could only be bootrstrapped <1000 times be included?}

\item{control_label}{How should the "control" group be labelled?}

\item{treatment_label}{How should the "treatment" group be labelled?}

\item{control_col}{Which colour should the "control" group bars be plotted in?}

\item{treatment_col}{Which colour should the "treatment" group bars be plotted in?}

\item{sort_by}{Should the bars be sorted by significance in the "treatment" or "control" group? (Default sort_by="treatment")}

\item{topN}{How many terms should be displayed? The terms are ordered by significance and only the top n terms plotted comparatively.}

\item{filepath}{Path + name of png to save If not provided, only the ggplot will be returned by this function.}
}
\value{
Returns a ggplot which can be stored in a variable and modified. If an absolute file name (path + file name) is provided, the ggplot with additionally be saved to a png.
}
\description{
Comparative visualisation of the GeneCOCOA results for a GOI in two different conditions.
Produces a ggplot divering bar plot with x=-log10(p.adj) and y= top significant terms in chosen condition
}
\examples{
\dontrun{

# load complete Familial Hypercholesteroleamia data set
FH <- as.data.frame(get_dataset_expression("GSE6054"))
FH_disease <- FH \%>\% select(contains("FH"))
disease_input <- get_expr_info(expr=FH_disease, GOI="LDLR")
disease_res <- get_stats(geneset_list=hallmark_sets,
                         GOI="LDLR",
                         GOI_expr=disease_input$GOI_expr,
                         expr_df=disease_input$expr_df,
                         samplesize=2, nsims=1000)


FH_control <- FH \%>\% select(contains("Control"))
control_input <- get_expr_info(expr=FH_control, GOI="LDLR")
control_res <- get_stats(geneset_list=hallmark_sets,
                         GOI="LDLR",
                         GOI_expr=control_input$GOI_expr,
                         expr_df=control_input$expr_df,
                         samplesize=2, nsims=1000)


p <- plot_control_vs_treatment(control_res, treatment_res,
                               control_label="control", treatment_label="FH",
                               topN=5, sort_by="treament",
                               filepath="/path/to/my/LDLR.FH.diverging_bars.png")
p
}
}
