% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot_differential_GeneCOCOA_results}
\alias{plot_differential_GeneCOCOA_results}
\title{Comparative visualisation of the GeneCOCOA results for a GOI in two different conditions.
Produces a ggplot divering bar plot with x=-log10(p.adj) and y= top significant terms in chosen condition}
\usage{
plot_differential_GeneCOCOA_results(
  diff_df,
  sig_label_cutoff = 0.05,
  output_dir = getwd(),
  GOI = "GOI",
  condition = ""
)
}
\arguments{
\item{diff_df}{Output of \code{\link{get_differential_GeneCOCOA_results}}}

\item{sig_label_cutoff}{At which differential p-value should points be labelled? Default: 0.05}

\item{output_dir}{Output path for the plots. Default: working directory.}

\item{GOI}{A string specifying the GOI gene symbol.}

\item{condition}{A string specifying the treatment condition examined in contrast to control.}
}
\value{
Returns a list with two ggplots, a MA plot wand a volcano plot. Additionally saves them to files.
}
\description{
Comparative visualisation of the GeneCOCOA results for a GOI in two different conditions.
Produces a ggplot divering bar plot with x=-log10(p.adj) and y= top significant terms in chosen condition
}
