% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{plot_multiple_p_plots}
\alias{plot_multiple_p_plots}
\title{Function to plot multiple (long) p-value rankings. Obsolete.}
\usage{
plot_multiple_p_plots(subdf)
}
\description{
Function to plot multiple (long) p-value rankings. Obsolete.
}
\keyword{Internal}
