% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneCOCOA.R
\name{rmse}
\alias{rmse}
\title{Returns root mean square error of a given regression model.}
\usage{
rmse(my_model)
}
\arguments{
\item{my_model}{}
}
\description{
Returns root mean square error of a given regression model.
}
\keyword{Internal}
